Moved the library files in lib
