+++ lib/index.js
@@ -1,0 +1,55 @@
+// Dependencies
+var Colors = require ("./colors");
+
+
+/**
+ *  Node Flat Colors
+ *
+ *  Convert any color to the nearest flat ui color.
+ *
+ * */
+function FlatColors(r, g, b) {
+
+    if (typeof r === "string" && r.charAt(0) === "#") {
+        return FlatColors(FlatColors.toRgb(r));
+    }
+
+    var rgb = Array.isArray(r) ? r : [ r, g, b ]
+      , best = null
+      ;
+
+    for (var i = 0; i < Colors.length; ++i) {
+        var d = distance(Colors[i], rgb)
+        if (!best || d <= best.distance) {
+            best = { distance : d, index : i };
+        }
+    }
+
+    return Colors[best.index];
+}
+
+FlatColors.toRgb = function (hex) {
+    var shorthandRegex = /^#?([a-f\d])([a-f\d])([a-f\d])$/i;
+    hex = hex.replace(shorthandRegex, function(m, r, g, b) {
+        return r + r + g + g + b + b;
+    });
+    var result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
+    return result ? [
+        parseInt(result[1], 16)
+      , parseInt(result[2], 16)
+      , parseInt(result[3], 16)
+    ] : null;
+};
+
+FlatColors.colors = Colors;
+
+
+function distance (a, b) {
+    return Math.sqrt(
+        Math.pow(a[0]-b[0], 2)
+        + Math.pow(a[1]-b[1], 2)
+        + Math.pow(a[2]-b[2], 2)
+    )
+}
+
+module.exports = FlatColors;
